Rename cms branch param from master -> main
Github has changed from using master -> main as the default branch for new repos.  So when netifly creates a new repo based on this template, the default branch is named 'main' instead of 'master'

Leaving this config param as 'master' will result in a credentials error whenever a user tries to login using the /admin page. Renaming this to 'main' fixes the error.